requirements/production.txt
# Dependencies related to deploying reformed-nous
-r base.txt

# WSGI Handler
gunicorn==19.3.0

# Caching
django-redis==4.4.3
